rewrite the crossdocking test
the previous version was not generating any PO because the product and
warehouse were not correctly configured. It did seem to work because
it found the PO from the cancelation_propagated.yml test and the final step
was actually trying to confirm a cancelled PO.

This new version actually tests the crossdocking flow.

Since this test has been rewritten, @nim-odoo updated the original one,
but still the generated move is not checked to be actually crossdock. We
keep both, just in case.