rewrite the crossdocking test
the previous version was not generating any PO because the product and
warehouse were not correctly configured. It did seem to work because
it found the PO from the cancelation_propagated.yml test and the final step
was actually trying to confirm a cancelled PO.

This new version actually tests the crossdocking flow.